# docs/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"paths":{"/v1/product":{"put":{"tags":["product-controller"],"description":"Update a product","operationId":"updateProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetProductResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Unit"}}}},"500":{"description":"Internal Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Unit"}}}}}},"post":{"tags":["product-controller"],"description":"Create a product","operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetProductResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Unit"}}}},"500":{"description":"Internal Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Unit"}}}}}},"delete":{"tags":["product-controller"],"description":"Delete a product by id","operationId":"deleteProduct","parameters":[],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetProductResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Unit"}}}},"500":{"description":"Internal Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Unit"}}}}}}},"/v1/product/{id}":{"get":{"tags":["product-controller"],"description":"Get a product by type","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetProductResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Unit"}}}},"500":{"description":"Internal Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Unit"}}}}}}},"/v1/product/type/{type}":{"get":{"tags":["product-controller"],"description":"Get a list of products by type","operationId":"getByType","parameters":[{"name":"type","in":"path","required":true,"schema":{"type":"string","enum":["MAIN","SIDE","BEVERAGE","DESSERT"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetProductResponse"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Unit"}}}},"500":{"description":"Internal Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Unit"}}}}}}}},"components":{"schemas":{"GetProductResponse":{"required":["id","name","price","type"],"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"price":{"type":"number","format":"double"},"image":{"type":"string"},"type":{"type":"string","enum":["MAIN","SIDE","BEVERAGE","DESSERT"]}}},"Unit":{"type":"object"},"UpdateProductRequest":{"required":["description","id","name","price","type"],"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"maxLength":200,"minLength":0,"type":"string"},"price":{"minimum":0,"type":"number","format":"double"},"image":{"type":"string"},"type":{"type":"string","enum":["MAIN","SIDE","BEVERAGE","DESSERT"]}}},"CreateProductRequest":{"required":["description","name","price","type"],"type":"object","properties":{"name":{"type":"string"},"description":{"maxLength":200,"minLength":0,"type":"string"},"price":{"minimum":0,"type":"number","format":"double"},"image":{"type":"string"},"type":{"type":"string","enum":["MAIN","SIDE","BEVERAGE","DESSERT"]}}}}}}